Move converters and enums in StaticDataEndPoint
Move it in different folders and namespaces in StaticDataEndPoint
